requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
requests==2.28.2
black==23.1.0
isort==5.12.0
ipdb==0.13.13
django-extensions==3.2.1